take modulo the seed on the server side when setting the cvar

+++ Quake/archipelago/apquake.cc
@@ -909,7 +909,7 @@ ap_level_index_t ap_make_level_index(int ep /* 1-based */, int map /* 1-based */
 	return { ep - 1, map - 1 };
 }
 
-float ap_get_rand_seed()
+unsigned long long ap_get_rand_seed()
 {
     const char *str = apquake_get_seed();
     unsigned long long hash = 5381;
@@ -918,7 +918,7 @@ float ap_get_rand_seed()
     while ((c = *str++))
         hash = ((hash << 5) + hash) + c; /* hash * 33 + c */
 
-    return (float) hash;
+    return hash;
 }
 
 std::string mask_text(std::string text) {
